Add RightMediaCircle element across Anthem and Ember themes
Introduced `RightMediaCircle` element with necessary logic and configurations for both Anthem and Ember themes. Refactored `targetImageSize` method to enhance reusability for image dimension adjustments. Updated respective element factories and blocks kit to integrate the new element seamlessly.
